backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adk-extra-services>=0.1.5",
    "bcrypt>=4.2.1",
    "boto3>=1.40.22",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "google-adk>=1.13.0",
    "litellm>=1.76.1",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "SQLAlchemy>=2.0.32",
    "aiosqlite>=0.20.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "uvicorn>=0.35.0",
]
